read + verify blockstates in a region's chunks
Blockstates are validated in each section in a chunk by checking that
the blockstates are a valid index in the palette, so it's kinda basic
and doesn't even check if the blockstates aren't all just zero.

All of the structs and stuff set up to make the sketchy test work should
be a good enough foundation to add more stuff to, like actually parsing
the palette and then sending chunks off to the client.