read chunks and send them to the client
Instead of generating a layer of cobblestone and sending it, chunks are
read from a region file and sent to the client.

It's still pretty hacky and broken, though; chunks are read inside the
chunk_data protocol function instead of being passed in, and even though
the client is properly receiving these chunks, every block is air.
Probably something wrong with write_blockstate and network_blockstates
but idk