leave blockstates in compressed form
Leaving them compressed across longs uses less memory and maybe makes
sending them faster because they don't have to be re-compacted every
time they're sent, but idk what I'm talking about.

The read + write functions are pretty simple, and it removes a layer of
junk, so I think it's nice. The only bad thing I see with this approach
is that the blockstates array needs to be re-made whenever
bits_per_block changes, but that probably won't happen very often so I
think it's worth it :)