nbt: fix seeking for a TAG_End stopping at the first one it finds
Seeking for a TAG_End with nbt_tag_seek_iter() would stop at the first
TAG_End it finds, regardless of if that TAG_End corresponds to the root
compound or a child of the root compound. Now it just skips over lists
of compound nodes if searching for a TAG_End.

With this fix, a lot more chunk sections are being parsed correctly (but
there's still some holes here and there)