add default block IDs with and without properties in their name
Block IDs in the palettes in Anvil files seem to be split half and half
between having no properties for the default blockstate and having the
default property values for the default block state (ex. the default
state for water is sometimes "minecraft:water" and sometimes
"minecraft:water;level=0"). So, when parsing a default state in
blocks.json, just add both to the table.