Replace hand-written protocol functions with generated code
add slot read/write functions

handle reading and writing floats and doubles

probably just gonna move the packet code to packet-auto-gen
cus there's no point having it here anymore and it's a hassle
for the packet generation code to access it as-is, it needs to
know where chowder is cloned to and that's stupid

use the generated read/write functions instead of old protocol functions

This compiles, but it shouldn't work. See the FIXME in server.c

add and use functions for one-off packet read/writes

The old functions in protocol.c already called read_header() and
write_packet() on their own, but the autogenerated functions don't
so some extra wrapper functions are needed for that.

fix not all chunks being sent