world: reimplement world_get_chunks() the easy way
The previous implementation was painful to read because I thought
it would be faster (without actually benchmarking anything),
and it didn't even work for arbitrary view distances. The new
implementation is actually readable and doesn't have the view
distance issue. There's definitely room for improvement, but
at least it's readable now :)