include packet-auto-gen in the repo rather than as a submodule
Hosting packet-auto-gen in a seperate repo was more hassle than
it's worth. The server and packet-auto-gen depend too much on
eachother: tests in packet-auto-gen depend on files in the main
repo, and changes are usually made to both projects at the same
time which causes double-commits. When branches are involved,
there has to be a branch in both repos, which is kinda annoying.
So, merge them :)