packet-auto-gen: search the current scope during field-name resolution
The previous field-search implementation was dumb and just
searched the whole field-tree for a match, regardless of what
'scope' that field was in. For example, in the "Update Light"
packet [1], the 'length' field in the block-light-array struct
would reference the length field in the sky-light-array struct
because of how the old implementation worked. Now, it will
correctly find the field in the block-light-array struct.

I should probably add a test case for the "Update Light"
packet to show this, but I'm tired right now xd

[1] https://wiki.vg/index.php?title=Protocol&oldid=16067#Update_Light